+++ activesupport/lib/active_support/core_ext/date_time/calculations.rb
@@ -132,35 +132,47 @@ class DateTime
   alias :at_noon :middle_of_day
   alias :at_middle_of_day :middle_of_day
 
-  # Returns a new DateTime representing the end of the day (23:59:59).
+  # Returns a new DateTime representing the end of the day (23:59:59.999999).
   def end_of_day
     change(hour: 23, min: 59, sec: 59, usec: Rational(999999999, 1000))
   end
   alias :at_end_of_day :end_of_day
 
-  # Returns a new DateTime representing the start of the hour (hh:00:00).
+  # Returns a new DateTime representing the start of the hour (hh:00:00.000000).
   def beginning_of_hour
     change(min: 0)
   end
   alias :at_beginning_of_hour :beginning_of_hour
 
-  # Returns a new DateTime representing the end of the hour (hh:59:59).
+  # Returns a new DateTime representing the end of the hour (hh:59:59.999999).
   def end_of_hour
     change(min: 59, sec: 59, usec: Rational(999999999, 1000))
   end
   alias :at_end_of_hour :end_of_hour
 
-  # Returns a new DateTime representing the start of the minute (hh:mm:00).
+  # Returns a new DateTime representing the start of the minute (hh:mm:00.000000).
   def beginning_of_minute
     change(sec: 0)
   end
   alias :at_beginning_of_minute :beginning_of_minute
 
-  # Returns a new DateTime representing the end of the minute (hh:mm:59).
+  # Returns a new DateTime representing the end of the minute (hh:mm:59.999999).
   def end_of_minute
     change(sec: 59, usec: Rational(999999999, 1000))
   end
   alias :at_end_of_minute :end_of_minute
+
+  # Returns a new DateTime representing the start of the second (hh:mm:ss.000000).
+  def beginning_of_second
+    change(sec: 0)
+  end
+  alias :at_beginning_of_second :beginning_of_second
+
+  # Returns a new DateTime representing the end of the second (hh:mm:ss.999999).
+  def end_of_second
+    change(sec: 59, usec: Rational(999999999, 1000))
+  end
+  alias :at_end_of_second :end_of_second
 
   # Returns a <tt>Time</tt> instance of the simultaneous time in the system timezone.
   def localtime(utc_offset = nil)